requirements.txt
Flask==3.1.0
gunicorn
psycopg2-binary
click==8.1.8
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
